+++ tests/testthat/test-cfbd_draft_positions.R
@@ -1,0 +1,13 @@
+context("NFL Draft Positions")
+
+
+cols <- c(
+  "position_name", "position_abbreviation"
+)
+
+test_that("NFL Draft Positions", {
+  skip_on_cran()
+  x <- cfbd_draft_teams()
+  expect_equal(colnames(x), cols)
+  expect_s3_class(x, "data.frame")
+})